+++ example-classification-cm-demoapp.yaml
@@ -8,19 +8,15 @@ metadata:
 data:
   default.classification.yaml: |
     # Example of Flow Classification rules file, showcasing using "extractors" and raw rego.
-
     selector:
       service: demo1-demo-app.demoapp.svc.cluster.local
       control_point: { traffic: ingress }
-
     rules:
       # An example rule using extractor.
       # See following RFC for list of available extractors and their syntax.
-      # https://www.notion.so/YAML-classification-policies-extractors-a7f36469fdf14c1bbc79c6ba963680a5#9ae103e0c9034c4aba30228d05d91391
       ua:
         extractor:
           from: request.http.headers.user-agent
-
       # The same rule using raw rego. Requires specifying rego source code and a query
       also-ua:
         rego:
@@ -29,7 +25,6 @@ data:
             import input.attributes.request.http
             ua = http.headers["user-agent"]
           query: data.my.rego.pkg.ua
-
       user-type:
         extractor:
           from: request.http.headers.user-type
