+++ pkg/flowcontrol/common/flowcontrol.go
@@ -1,4 +1,4 @@
-package flowcontrol
+package common
 
 import (
 	"context"
