+++ pkg/flowcontrol/common/metrics.go
@@ -1,4 +1,4 @@
-package flowcontrol
+package common
 
 import (
 	"github.com/prometheus/client_golang/prometheus"
